+++ app/Filament/Resources/AdminActivityLogs/Pages/ViewAdminActivityLog.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\AdminActivityLogs\Pages;
+
+use App\Filament\Resources\AdminActivityLogs\AdminActivityLogResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewAdminActivityLog extends ViewRecord
+{
+    protected static string $resource = AdminActivityLogResource::class;
+}